Code refactor and new test cases
