--- README.md
@@ -1,2 +1,0 @@
-# Budgety
-Simple app for budget management
