+++ settings.gradle.kts
@@ -1,15 +1,1 @@
-pluginManagement {
-	repositories {
-		maven { url = uri("https://repo.spring.io/milestone") }
-		maven { url = uri("https://repo.spring.io/snapshot") }
-		gradlePluginPortal()
-	}
-	resolutionStrategy {
-		eachPlugin {
-			if (requested.id.id == "org.springframework.boot") {
-				useModule("org.springframework.boot:spring-boot-gradle-plugin:${requested.version}")
-			}
-		}
-	}
-}
-rootProject.name = "budgety"
+rootProject.name = "budget"
