+++ src/main/kotlin/com/budget/BudgetApplication.kt
@@ -1,0 +1,11 @@
+package com.budget
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class BudgetApplication
+
+fun main(args: Array<String>) {
+	runApplication<BudgetApplication>(*args)
+}
