--- src/main/kotlin/com/budgety/ServiceApplication.kt
@@ -1,16 +1,0 @@
-package com.budgety
-
-import com.budgety.adapter.dao.CategoryDAO
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.autoconfigure.data.mongo.MongoDataAutoConfiguration
-import org.springframework.boot.autoconfigure.mongo.MongoAutoConfiguration
-import org.springframework.boot.runApplication
-import org.springframework.data.mongodb.repository.config.EnableMongoRepositories
-
-@SpringBootApplication(exclude = [MongoAutoConfiguration::class, MongoDataAutoConfiguration::class])
-@EnableMongoRepositories(basePackages = ["com.budgety.adapter.dao"])
-class ServiceApplication
-
-fun main(args: Array<String>) {
-	runApplication<ServiceApplication>(*args)
-}
