--- src/main/kotlin/com/budgety/dto/CategoryDto.kt
@@ -1,5 +1,0 @@
-package com.budgety.dto
-
-data class CategoryDto(val code: String, val name: String){
-    var description : String = ""
-}
