--- src/main/kotlin/com/budgety/service/model/Category.kt
@@ -1,5 +1,0 @@
-package com.budgety.service.model
-
-data class Category(val code: String,val name: String){
-    var description : String = ""
-}
