+++ src/main/resources/application.properties
@@ -1,8 +1,1 @@
-spring.data.mongodb.authentication-database=admin
-spring.data.mongodb.database=budgety
-#replace with your mongodb host
-spring.data.mongodb.host=budgety-alatv.mongodb.net
-spring.data.mongodb.port=27017
-#replace with your username & password
-spring.data.mongodb.username=admin
-spring.data.mongodb.password=admin_pswd
+
