+++ src/test/kotlin/com/budget/BudgetApplicationTests.kt
@@ -1,0 +1,13 @@
+package com.budget
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class BudgetApplicationTests {
+
+	@Test
+	fun contextLoads() {
+	}
+
+}
